+++ cert_viewer/templates/layout.html
