+++ cert_viewer/verification_helpers.py
