+++ conf_template.ini
@@ -16,8 +16,8 @@ CRITERIA_PATH = <criteria/>
 KEYS_PATH = 'keys/'
 CERT_PUBKEY = <public-issuing-key>
 CERT_REVOKEKEY = <public-revocation-key>
-ISSUER_PATH = 'issuer/'
-MLISSUER_PATH = ISSUER_PATH + 'ml-issuer.json'
+ISSUER_DIR = 'issuer/'
+ISSUER_PATH = ISSUER_DIR + 'ml-issuer.json'
 
 [logging]
 LOG_DIR = /tmp/
