+++ docs/index.rst
@@ -1,4 +1,4 @@
-.. viewer documentation master file, created by
+.. cert-viewer documentation master file, created by
    sphinx-quickstart on Tue May 24 12:06:50 2016.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
@@ -10,7 +10,7 @@ contributions, and general discussion. This is not currently intended for produc
 our approach for future releases. Read more about the Digital Certificates Project at http://certificates.media.mit.edu/.
 
 
-Get the code: https://github.com/digital-certificates/viewer
+Get the code: https://github.com/digital-certificates/cert-viewer
 
 
 Getting started
@@ -52,15 +52,6 @@ API docs
    :maxdepth: 2
 
    modules
-
-
-
-Indices and tables
-------------------
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
 
 
 Contact
