+++ docs/installation.md
@@ -8,7 +8,7 @@ should be changed if you're using this for anything other than demo purposes.
 1. Git clone the repository
 
     ```bash
-    git clone https://github.com/digital-certificates/viewer.git
+    git clone https://github.com/digital-certificates/cert-viewer.git
     ```
 
 2. Install [mongodb](https://docs.mongodb.com/v3.0/installation/)
@@ -16,7 +16,7 @@ should be changed if you're using this for anything other than demo purposes.
 3. Set up a virtual environment [details](http://docs.python-guide.org/en/latest/dev/virtualenvs/) and run setup
 
     ```bash
-    cd viewer
+    cd cert-viewer
     pip install .
     ```
