+++ docs/quick_start.md
@@ -10,7 +10,7 @@ example, mac instructions start at [https://docs.docker.com/engine/installation/
 
 2. Git clone the repository
 
-    `git clone https://github.com/digital-certificates/viewer.git`
+    `git clone https://github.com/digital-certificates/cert-viewer.git`
 
 3. Build the container with docker-compose
 
@@ -20,7 +20,7 @@ example, mac instructions start at [https://docs.docker.com/engine/installation/
 
     `docker-compose up`
 
-5. Access the certificate-viewer pre-populated with test data at `http://<hostname>:5000`, where hostname is given by
+5. Access the cert-viewer pre-populated with test data at `http://<hostname>:5000`, where hostname is given by
     `docker-machine ip`
 
 
@@ -28,7 +28,7 @@ About Docker Setup
 ------------------
 The quick start steps do the following:
 
-1. Creates a container that runs the certificate-viewer Flask app with MongoDB using Docker Compose [details](http://containertutorials.com/docker-compose/flask-mongo-compose.html)
+1. Creates a container that runs the cert-viewer Flask app with MongoDB using Docker Compose [details](http://containertutorials.com/docker-compose/flask-mongo-compose.html)
 2. Seeds the MongoDB database with sample fake certificates. This data is located in the mongo-seed folder
 3. Starts the container. This configuration exposes port 5000.
