+++ run.py
@@ -1,2 +1,2 @@
-from viewer import app
-app.run(host='0.0.0.0', debug=True)
+from cert_viewer import app
+app.run(host='0.0.0.0', port=5000, debug=True)
